mibot/requirements.txt
art
beautifulsoup4
bs4
googletrans
gTTS
html5lib
humanfriendly
humanize
livejson
lxml
Naked
pafy
pillow
PyQRCode
python-axolotl-curve25519
pytz
pytube
requests
requests_html
rsa
schedule
six
spotdl
thrift
timeago
youtube-dl
youtube-search-python
